+++ heating_emissions/components/histogram_artifacts.py
@@ -4,20 +4,33 @@ import plotly.graph_objects as go
 from climatoology.base.artifact import Artifact, ArtifactMetadata
 from climatoology.base.artifact_creators import create_plotly_chart_artifact
 from climatoology.base.computation import ComputationResources
+from climatoology.base.i18n import tr
 from plotly.graph_objs import Figure
 
 from heating_emissions.components.utils import Topics
 
 
+def format_summary(summary: str, figure: Figure) -> str:
+    return summary.format(
+        mean_data=round(np.mean(figure['data'][0].x), 2),
+        min_data=round(np.min(figure['data'][0].x), 2),
+        max_data=round(np.max(figure['data'][0].x), 2),
+    )
+
+
 def build_per_capita_direct_co2_histogram_artifact(aoi_aggregate: Figure, resources: ComputationResources) -> Artifact:
+    summary = tr(
+        'Average direct carbon dioxide emissions from heating residential buildings (estimated) '
+        'are {mean_data} tonnes per person per year, '
+        'but range from {min_data} to '
+        ' {max_data} tonnes.'
+    )
+    summary = format_summary(summary, aoi_aggregate)
     per_capita_direct_co2_histogram_artifact_metadata = ArtifactMetadata(
-        name='Histogram per capita direct emissions',
-        summary=f'Average direct carbon dioxide emissions from heating residential buildings (Estimated) '
-        f'are {round(np.mean(aoi_aggregate["data"][0].x), 2)} tonnes per person per year, '
-        f'but range from {round(np.min(aoi_aggregate["data"][0].x), 2)} to '
-        f' {round(np.max(aoi_aggregate["data"][0].x), 2)} tonnes.',
+        name=tr('Histogram of per capita direct emissions'),
+        summary=summary,
         filename='aoi_direct_C02pc_histogram',
-        tags={Topics.DIRECT_EMISSIONS},
+        tags={tr(Topics.DIRECT_EMISSIONS)},
     )
     return create_plotly_chart_artifact(
         figure=aoi_aggregate,
@@ -29,14 +42,18 @@ def build_per_capita_direct_co2_histogram_artifact(aoi_aggregate: Figure, resour
 def build_per_capita_life_cycle_co2_histogram_artifact(
     aoi_aggregate: Figure, resources: ComputationResources
 ) -> Artifact:
+    summary = tr(
+        'Average life cycle GHG emissions from heating residential buildings (estimated) '
+        'are {mean_data} tonnes of CO₂-eq. per person per year, '
+        'but range from {min_data} to '
+        ' {max_data} tonnes.'
+    )
+    summary = format_summary(summary, aoi_aggregate)
     per_capita_life_cycle_co2_histogram_artifact_metadata = ArtifactMetadata(
-        name='Histogram per capita life cycle emissions ',
-        summary=f'Average life cycle GHG emissions from heating residential buildings (Estimated) '
-        f'are {round(np.mean(aoi_aggregate["data"][0].x), 2)} tonnes of CO₂-eq. per person per year, '
-        f'but range from {round(np.min(aoi_aggregate["data"][0].x), 2)} to '
-        f' {round(np.max(aoi_aggregate["data"][0].x), 2)} tonnes.',
+        name=tr('Histogram of per capita life cycle emissions'),
+        summary=summary,
         filename='aoi_life_cycle_C02pc_histogram',
-        tags={Topics.LIFE_CYCLE_EMISSIONS},
+        tags={tr(Topics.LIFE_CYCLE_EMISSIONS)},
     )
     return create_plotly_chart_artifact(
         figure=aoi_aggregate,
@@ -46,14 +63,18 @@ def build_per_capita_life_cycle_co2_histogram_artifact(
 
 
 def build_energy_histogram_artifact(aoi_aggregate: Figure, resources: ComputationResources) -> Artifact:
+    summary = tr(
+        'Average heating energy consumption in residential buildings (estimated) '
+        'is {mean_data} kWh per square meter per year, '
+        'but range from {min_data} to '
+        ' {max_data} .'
+    )
+    summary = format_summary(summary, aoi_aggregate)
     energy_histogram_artifact_metadata = ArtifactMetadata(
-        name='Histogram energy consumption',
-        summary=f'Average heating energy consumption in residential buildings (Estimated) '
-        f'is {round(np.mean(aoi_aggregate["data"][0].x), 2)} kWh per square meter per year, '
-        f'but range from {round(np.min(aoi_aggregate["data"][0].x), 2)} to '
-        f' {round(np.max(aoi_aggregate["data"][0].x), 2)} .',
+        name=tr('Histogram of energy consumption'),
+        summary=summary,
         filename='aoi_energy_histogram',
-        tags={Topics.PARAMETERS},
+        tags={tr(Topics.PARAMETERS)},
     )
     return create_plotly_chart_artifact(
         figure=aoi_aggregate,
@@ -63,14 +84,18 @@ def build_energy_histogram_artifact(aoi_aggregate: Figure, resources: Computatio
 
 
 def build_direct_emission_factor_histogram_artifact(aoi_aggregate: Figure, resources: ComputationResources) -> Artifact:
+    summary = tr(
+        'Average direct emission factor from heating residential buildings (estimated) '
+        'is {mean_data} kg of carbon dioxide per kWh, '
+        'but range from {min_data} to '
+        ' {max_data}.'
+    )
+    summary = format_summary(summary, aoi_aggregate)
     direct_emission_factor_histogram_artifact_metadata = ArtifactMetadata(
-        name='Histogram direct emission factor',
-        summary=f'Average direct emission factor from heating residential buildings (Estimated) '
-        f'is {round(np.mean(aoi_aggregate["data"][0].x), 2)} kg of carbon dioxide per kWh, '
-        f'but range from {round(np.min(aoi_aggregate["data"][0].x), 2)} to '
-        f' {round(np.max(aoi_aggregate["data"][0].x), 2)}.',
+        name=tr('Histogram of direct emission factor'),
+        summary=summary,
         filename='aoi_direct_ef_histogram',
-        tags={Topics.DIRECT_EMISSIONS},
+        tags={tr(Topics.DIRECT_EMISSIONS)},
     )
     return create_plotly_chart_artifact(
         figure=aoi_aggregate,
@@ -82,14 +107,18 @@ def build_direct_emission_factor_histogram_artifact(aoi_aggregate: Figure, resou
 def build_life_cycle_emission_factor_histogram_artifact(
     aoi_aggregate: Figure, resources: ComputationResources
 ) -> Artifact:
+    summary = tr(
+        'Average life cycle emission factor from heating residential buildings (estimated) '
+        'is {mean_data} kg CO₂-eq. per kWh, '
+        'but range from {min_data} to '
+        ' {max_data}.'
+    )
+    summary = format_summary(summary, aoi_aggregate)
     life_cycle_emission_factor_histogram_artifact_metadata = ArtifactMetadata(
-        name='Histogram life cycle emission factor',
-        summary=f'Average life cycle emission factor from heating residential buildings (Estimated) '
-        f'is {round(np.mean(aoi_aggregate["data"][0].x), 2)} kg CO₂-eq. per kWh, '
-        f'but range from {round(np.min(aoi_aggregate["data"][0].x), 2)} to '
-        f' {round(np.max(aoi_aggregate["data"][0].x), 2)}.',
+        name=tr('Histogram of life cycle emission factor'),
+        summary=summary,
         filename='aoi_life_cycle_ef_histogram',
-        tags={Topics.LIFE_CYCLE_EMISSIONS},
+        tags={tr(Topics.LIFE_CYCLE_EMISSIONS)},
     )
     return create_plotly_chart_artifact(
         figure=aoi_aggregate,
@@ -104,23 +133,25 @@ def plot_per_capita_direct_co2_histogram(
     hist_plot = Figure(
         data=go.Histogram(
             x=census_data['direct_co2_emissions_per_capita'] / 1e3,  # result in tonnes,
-            hovertemplate='%{y:.1f} % of cells emit %{x} tonnes of carbon dioxide per person per year <extra></extra>',
+            hovertemplate=tr(
+                '%{y:.1f} % of cells emit %{x} tonnes of carbon dioxide per person per year <extra></extra>'
+            ),
             histnorm='percent',
         ),
         layout=go.Layout(
             title=dict(
-                text='Per capita direct carbon dioxide emissions from residential heating (Estimated)',
-                subtitle=dict(text='% of 100-m grid cells in area of interest', font=dict(color='gray', size=10)),
+                text=tr('Per capita direct carbon dioxide emissions from residential heating (estimated)'),
+                subtitle=dict(text=tr('% of 100-m grid cells in area of interest'), font=dict(color='gray', size=10)),
             ),
             font=dict(size=12),
-            xaxis=dict(title=dict(text='annual tonnes of carbon dioxide per person')),
+            xaxis=dict(title=dict(text=tr('annual tonnes of carbon dioxide per person'))),
         ),
     )
     hist_plot.add_vline(
         x=2.2,
         line_width=1.5,
         line_dash='dot',
-        annotation_text='German average',
+        annotation_text=tr('German average'),
         annotation_position='top left',
         annotation_textangle=-90,
     )
@@ -133,23 +164,23 @@ def plot_per_capita_life_cycle_co2_histogram(
     hist_plot = Figure(
         data=go.Histogram(
             x=census_data['life_cycle_co2_emissions_per_capita'] / 1e3,  # result in tonnes,
-            hovertemplate='%{y:.1f} % of cells emit %{x} tonnes of kg CO₂-eq. per person per year <extra></extra>',
+            hovertemplate=tr('%{y:.1f} % of cells emit %{x} tonnes of kg CO₂-eq. per person per year <extra></extra>'),
             histnorm='percent',
         ),
         layout=go.Layout(
             title=dict(
-                text='Per capita life cycle GHG emissions from residential heating (Estimated)',
-                subtitle=dict(text='% of 100-m grid cells in area of interest', font=dict(color='gray', size=10)),
+                text=tr('Per capita life cycle GHG emissions from residential heating (estimated)'),
+                subtitle=dict(text=tr('% of 100-m grid cells in area of interest'), font=dict(color='gray', size=10)),
             ),
             font=dict(size=12),
-            xaxis=dict(title=dict(text='annual tonnes of CO₂-eq. per person')),
+            xaxis=dict(title=dict(text=tr('annual tonnes of CO₂-eq. per person'))),
         ),
     )
     hist_plot.add_vline(
         x=2.2,
         line_width=1.5,
         line_dash='dot',
-        annotation_text='German average',
+        annotation_text=tr('German average'),
         annotation_position='top left',
         annotation_textangle=-90,
     )
@@ -162,24 +193,26 @@ def plot_energy_consumption_histogram(
     hist_plot = Figure(
         data=go.Histogram(
             x=census_data['heat_consumption'],
-            hovertemplate='Buildings in %{y:.1f} % of cells consume %{x} kWh of heating energy'
-            ' per square meter per year <extra></extra>',
+            hovertemplate=tr(
+                'Buildings in %{y:.1f} % of cells consume %{x} kWh of heating energy'
+                ' per square meter per year <extra></extra>'
+            ),
             histnorm='percent',
         ),
         layout=go.Layout(
             title=dict(
-                text='Average heating energy consumption rate in residential buildings (Estimated)',
-                subtitle=dict(text='% of 100-m grid cells in area of interest', font=dict(color='gray', size=10)),
+                text=tr('Average area-specific heating energy consumption in residential buildings (estimated)'),
+                subtitle=dict(text=tr('% of 100-m grid cells in area of interest'), font=dict(color='gray', size=10)),
             ),
             font=dict(size=12),
-            xaxis=dict(title=dict(text='kWh per square meter per year')),
+            xaxis=dict(title=dict(text=tr('kWh per square meter per year'))),
         ),
     )
     hist_plot.add_vline(
         x=127.1,
         line_width=1.5,
         line_dash='dot',
-        annotation_text='German average',
+        annotation_text=tr('German average'),
         annotation_position='top left',
         annotation_textangle=-90,
     )
@@ -192,24 +225,26 @@ def plot_direct_emission_factor_histogram(
     hist_plot = Figure(
         data=go.Histogram(
             x=census_data['direct_emission_factor'],
-            hovertemplate='Buildings in %{y:.1f} % of cells emit %{x} kg of carbon dioxide'
-            ' per kWh of heating energy used <extra></extra>',
+            hovertemplate=tr(
+                'Buildings in %{y:.1f} % of cells emit %{x} kg of carbon dioxide'
+                ' per kWh of heating energy used <extra></extra>'
+            ),
             histnorm='percent',
         ),
         layout=go.Layout(
             title=dict(
-                text='Average direct emission factor from heating in residential buildings (Estimated)',
-                subtitle=dict(text='% of 100-m grid cells in area of interest', font=dict(color='gray', size=10)),
+                text=tr('Average direct emission factor from heating in residential buildings (estimated)'),
+                subtitle=dict(text=tr('% of 100-m grid cells in area of interest'), font=dict(color='gray', size=10)),
             ),
             font=dict(size=12),
-            xaxis=dict(title=dict(text='kg of carbon dioxide per kWh of heating')),
+            xaxis=dict(title=dict(text=tr('kg of carbon dioxide per kWh of heating'))),
         ),
     )
     hist_plot.add_vline(
         x=0.199,
         line_width=1.5,
         line_dash='dot',
-        annotation_text='German average',
+        annotation_text=tr('German average'),
         annotation_position='top left',
         annotation_textangle=-90,
     )
@@ -222,24 +257,26 @@ def plot_life_cycle_emission_factor_histogram(
     hist_plot = Figure(
         data=go.Histogram(
             x=census_data['life_cycle_emission_factor'],
-            hovertemplate='Buildings in %{y:.1f} % of cells emit %{x} kg of carbon dioxide'
-            ' per kWh of heating energy used <extra></extra>',
+            hovertemplate=tr(
+                'Buildings in %{y:.1f} % of cells emit %{x} kg of carbon dioxide'
+                ' per kWh of heating energy used <extra></extra>'
+            ),
             histnorm='percent',
         ),
         layout=go.Layout(
             title=dict(
-                text='Average life cycle emission factor from heating in residential buildings (Estimated)',
-                subtitle=dict(text='% of 100-m grid cells in area of interest', font=dict(color='gray', size=10)),
+                text=tr('Average life cycle emission factor from heating in residential buildings (estimated)'),
+                subtitle=dict(text=tr('% of 100-m grid cells in area of interest'), font=dict(color='gray', size=10)),
             ),
             font=dict(size=12),
-            xaxis=dict(title=dict(text='kg of carbon dioxide per kWh of heating')),
+            xaxis=dict(title=dict(text=tr('kg of carbon dioxide per kWh of heating'))),
         ),
     )
     hist_plot.add_vline(
         x=0.199,
         line_width=1.5,
         line_dash='dot',
-        annotation_text='German average',
+        annotation_text=tr('German average'),
         annotation_position='top left',
         annotation_textangle=-90,
     )