+++ heating_emissions/core/input.py
@@ -1,6 +1,7 @@
 from datetime import datetime
 from typing import Annotated, Optional
 
+from climatoology.base.i18n import N_
 from pydantic import BaseModel, Field
 from pydantic.json_schema import SkipJsonSchema
 
@@ -13,12 +14,14 @@ class ComputeInput(BaseModel):
     temporal_emission_year: Annotated[
         Optional[int],
         Field(
-            title='Temporal Downscaling',
-            description='Specify the year to be used for temporal flexible simulation of heating emissions. '
-            'If `None` (the default) then do not apply temporal downscaling. '
-            'The emissions are calculated based on '
-            'simulated heating demand profiles derived from ERA5 weather data. '
-            'Warning: The download of the ERA5 weather data may take a long time. ',
+            title=N_('Temporal Downscaling'),
+            description=N_(
+                'Specify the year to be used for temporal flexible simulation of heating emissions. '
+                'If `None` (the default) then do not apply temporal downscaling. '
+                'The emissions are calculated based on '
+                'simulated heating demand profiles derived from ERA5 weather data. '
+                'Warning: The download of the ERA5 weather data may take a long time. '
+            ),
             ge=2017,
             le=datetime.now().year - 1,
             examples=[2022],