+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 - Clarify the data year '2022' & add scope 1 & life-cycle in purpose.md ([#63](https://gitlab.heigit.org/climate-action/plugins/heating-emissions/-/work_items/63))
 - Add how this plugin relates to climate action in purpose.md ([#76](https://gitlab.heigit.org/climate-action/plugins/heating-emissions/-/work_items/76))
 - Add cdsapi_client authentication check ([#78](https://gitlab.heigit.org/climate-action/plugins/heating-emissions/-/work_items/78))
+- Translation to German ([#65](https://gitlab.heigit.org/climate-action/plugins/heating-emissions/-/work_items/65))
 
 ### Changed